requirements.txt
click>=8.1
gitpython>=3.1
rich>=13.0
tree-sitter>=0.21
tree-sitter-python>=0.21
tree-sitter-javascript>=0.21
tree-sitter-typescript>=0.21
chromadb>=0.4
sentence-transformers>=2.2
groq>=0.4
requests>=2.28
